common: add Upstart task for setting boot completion property.
This script sets a number of Android properties which indicate if the
boot process has fully completed. Qualcomm's init system depends on
this property to carry some action.

For example, on Fairphone 2, this triggers init.qcom.post_boot.sh which
set a number of low power mode for CPU. This resulted in lower power
consumption and heat.