(3/3) Make TimeKeep available to keep time offset
TimeKeep is utility to keep/restore RTC offset for Qualcomm device, where
RTC is not setable. It works by keeping the offset between the RTC time and
system time in a property (or a file) and restore it at boot.

This commit add TimeKeep's unit file so that the time is restored at boot
and is saved at shutdown. If any device doesn't have TimeKeep binary, it'll
silently skip this process.

Change-Id: I090d14ed6ad969ffef010cf2756eb1e710ca87ce